Update VS Dashboard chart labels and tick numbers
