+++ frontend/src/pages/VulnerabilityScanDash/SeverityByProminence.tsx
@@ -1,6 +1,12 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useTheme } from '@mui/material/styles';
+import { Box, Grid } from '@mui/material';
 import { axisClasses, BarChart } from '@mui/x-charts';
+import GraphChip from 'components/Dashboard/GraphChip';
+import InfoLabel from 'components/Dashboard/InfoLabel';
+import infoIconContent from './infoIconContent.json';
+
+const tooltipContentJson = infoIconContent.infoIconContent;
 
 interface SeverityByProminenceProps {
   data: {
@@ -10,7 +16,6 @@ interface SeverityByProminenceProps {
     highSeverity: number;
     criticalSeverity: number;
   }[];
-  activeVulnType?: string;
 }
 
 const SevByPromBarChart = (props: any) => {
@@ -18,9 +23,9 @@ const SevByPromBarChart = (props: any) => {
 };
 
 const SeverityByProminence: React.FC<SeverityByProminenceProps> = ({
-  data,
-  activeVulnType
+  data
 }) => {
+  const [activeLabel, setActiveLabel] = useState('KEV');
   const theme = useTheme();
   const gridData = data.map((item) => ({
     type: item.vulnType,
@@ -30,7 +35,16 @@ const SeverityByProminence: React.FC<SeverityByProminenceProps> = ({
     critical: item.criticalSeverity
   }));
 
-  const filteredData = gridData.find((item) => item.type === activeVulnType);
+  const handleSevByProminenceChipClick = (label: string) => {
+    setActiveLabel(label);
+  };
+  const severityByProminenceGraphChips = ['KEV', 'Distinct', 'All'].map(
+    (type) => ({
+      label: type,
+      onClick: () => handleSevByProminenceChipClick(type)
+    })
+  );
+  const filteredData = gridData.find((item) => item.type === activeLabel);
 
   const findColor = (activeVulnType: string) => {
     switch (activeVulnType) {
@@ -64,13 +78,13 @@ const SeverityByProminence: React.FC<SeverityByProminenceProps> = ({
       ]
     : [];
 
-  return (
+  const barChart = (
     <SevByPromBarChart
       series={[
         {
           dataKey: 'value',
           label: 'Severity Count',
-          color: findColor(activeVulnType || 'All')
+          color: findColor(activeLabel || 'All')
         }
       ]}
       dataset={dataset}
@@ -82,8 +96,7 @@ const SeverityByProminence: React.FC<SeverityByProminenceProps> = ({
           labelStyle: {
             fontSize: 12,
             fontStyle: 'italic',
-            fontWeight: 500,
-            color: theme.palette.neutrals.main
+            fontWeight: 500
           },
           tickPlacement: 'end',
           tickLabelPlacement: 'middle',
@@ -99,12 +112,18 @@ const SeverityByProminence: React.FC<SeverityByProminenceProps> = ({
           labelStyle: {
             fontSize: 12,
             fontStyle: 'italic',
-            fontWeight: 500,
-            color: theme.palette.neutrals.main
+            fontWeight: 500
           },
           tickPlacement: 'end',
           tickLabelPlacement: 'tick',
-          tickLabelStyle: { fontSize: 10 }
+          tickLabelStyle: { fontSize: 10 },
+          disableTicks: true,
+          valueFormatter: (value: number) => {
+            if (Number.isInteger(value)) {
+              return value.toString();
+            }
+            return '';
+          }
         }
       ]}
       height={430}
@@ -116,10 +135,46 @@ const SeverityByProminence: React.FC<SeverityByProminenceProps> = ({
         [`.${axisClasses.root}`]: {
           [`.${axisClasses.line}`]: {
             strokeWidth: 3
+          },
+          [`.${axisClasses.label}`]: {
+            fill: '#646566'
           }
         }
       }}
     />
+  );
+  return (
+    <Box
+      border="1px solid"
+      borderRadius="4px"
+      borderColor="neutrals.light"
+      sx={{ px: 3, pb: 3, pt: 1 }}
+      minHeight="550px"
+      // maxWidth="538px"
+    >
+      <Grid
+        container
+        alignItems="center"
+        justifyContent="space-between"
+        paddingBottom={2}
+      >
+        <Grid size={{ xs: 12, lg: 7 }}>
+          <InfoLabel
+            label="Severity by Prominence"
+            typographyVariant="h3"
+            headingLevel="h3"
+            tooltipContentJson={tooltipContentJson}
+          />
+        </Grid>
+        <Grid size={{ xs: 12, lg: 5 }} justifyItems={'flex-end'}>
+          <GraphChip
+            data={severityByProminenceGraphChips}
+            activeLabel={activeLabel}
+          />
+        </Grid>
+      </Grid>
+      {barChart}
+    </Box>
   );
 };
