+++ frontend/src/pages/VulnerabilityScanDash/TopVulnerableHosts.tsx
@@ -1,8 +1,14 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import { axisClasses, BarChart } from '@mui/x-charts';
 import { useTheme } from '@mui/material/styles';
 import { getSeverityLevelColorMap } from 'utils/severityLevelColorMap';
+import { axisClasses, BarChart } from '@mui/x-charts';
+import { Box, Grid } from '@mui/material';
+import GraphChip from 'components/Dashboard/GraphChip';
+import InfoLabel from 'components/Dashboard/InfoLabel';
+import infoIconContent from './infoIconContent.json';
+
+const tooltipContentJson = infoIconContent.infoIconContent;
 
 interface TopVulnerableHostsProps {
   data: {
@@ -14,15 +20,12 @@ interface TopVulnerableHostsProps {
     all: number;
     domainId: string;
   }[];
-  severityLevel?: string;
 }
 const TopVulnHostsBarChart = (props: any) => {
   return <BarChart {...props} />;
 };
-const TopVulnerableHosts: React.FC<TopVulnerableHostsProps> = ({
-  data,
-  severityLevel
-}) => {
+const TopVulnerableHosts: React.FC<TopVulnerableHostsProps> = ({ data }) => {
+  const [activeSeverityLevel, setActiveSeverityLevel] = useState('all');
   const theme = useTheme();
   const history = useHistory();
   const gridData = data.map((item) => ({
@@ -35,25 +38,63 @@ const TopVulnerableHosts: React.FC<TopVulnerableHostsProps> = ({
     domainId: item.domainId
   }));
 
+  const handleTopVulnHostsChipClick = (label: string) => {
+    setActiveSeverityLevel(label);
+  };
+  const presentSeverityLevels = Array.from(
+    new Set(
+      data.flatMap((item) =>
+        Object.entries(item)
+          .filter(
+            ([key, value]) =>
+              [
+                'all',
+                'lowSeverity',
+                'mediumSeverity',
+                'highSeverity',
+                'criticalSeverity'
+              ].includes(key) &&
+              typeof value === 'number' &&
+              value > 0
+          )
+          .map(([key]) => key)
+      )
+    )
+  )
+    .map((severity) => severity.replace('Severity', '').toLowerCase())
+    .sort((a: string, b: string) => {
+      const severityOrder = ['all', 'critical', 'high', 'medium', 'low'];
+      return severityOrder.indexOf(a) - severityOrder.indexOf(b);
+    });
   const filteredData = gridData.map((item) => ({
     hostName: item.hostName,
-    value: item[severityLevel as keyof typeof item] || 0, // Use the active severity level to filter
+    value: item[activeSeverityLevel as keyof typeof item] || 0, // Use the active severity level to filter
     domainId: item.domainId
   }));
 
   const severityLevelColorMap = getSeverityLevelColorMap(theme);
   const severityLevelColor =
-    severityLevelColorMap[severityLevel as keyof typeof severityLevelColorMap];
+    severityLevelColorMap[
+      activeSeverityLevel as keyof typeof severityLevelColorMap
+    ];
 
   const capitalizeFirstLetter = (string: string) => {
     return string.charAt(0).toUpperCase() + string.slice(1);
   };
-  return (
+
+  const topVulnerableHostsGraphChips = presentSeverityLevels.map(
+    (severity) => ({
+      label: severity,
+      onClick: () => handleTopVulnHostsChipClick(severity)
+    })
+  );
+
+  const barChart = (
     <TopVulnHostsBarChart
       series={[
         {
           dataKey: 'value',
-          label: capitalizeFirstLetter(severityLevel || ''),
+          label: capitalizeFirstLetter(activeSeverityLevel || ''),
           color: severityLevelColor
         }
       ]}
@@ -66,11 +107,16 @@ const TopVulnerableHosts: React.FC<TopVulnerableHostsProps> = ({
           labelStyle: {
             fontSize: 12,
             fontStyle: 'italic',
-            fontWeight: 500,
-            color: theme.palette.neutrals.main
+            fontWeight: 500
           },
           disableTicks: true,
-          tickLabelStyle: { fontSize: 10 }
+          tickLabelStyle: { fontSize: 10 },
+          valueFormatter: (value: number) => {
+            if (Number.isInteger(value)) {
+              return value.toString();
+            }
+            return '';
+          }
         }
       ]}
       yAxis={[
@@ -144,10 +190,36 @@ const TopVulnerableHosts: React.FC<TopVulnerableHostsProps> = ({
         [`.${axisClasses.root}`]: {
           [`.${axisClasses.line}`]: {
             strokeWidth: 3
+          },
+          [`.${axisClasses.label}`]: {
+            fill: '#646566'
           }
         }
       }}
     />
+  );
+  return (
+    <Box
+      border="1px solid"
+      borderRadius="4px"
+      borderColor="neutrals.light"
+      sx={{ px: 3, pb: 2.5, pt: 1 }}
+      minHeight="435px"
+    >
+      <InfoLabel
+        label="Top Vulnerable Hosts"
+        typographyVariant="h3"
+        headingLevel="h3"
+        tooltipContentJson={tooltipContentJson}
+      />
+      <Grid container direction="row" paddingBottom={2}>
+        <GraphChip
+          data={topVulnerableHostsGraphChips}
+          activeLabel={activeSeverityLevel}
+        />
+      </Grid>
+      {barChart}
+    </Box>
   );
 };
