+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -10,7 +10,6 @@ import {
 import KeyMetrics from './KeyMetrics';
 import ScanningSummary from './ScanningSummary';
 import InfoLabel from 'components/Dashboard/InfoLabel';
-import GraphChip from 'components/Dashboard/GraphChip';
 import {
   // StatsTrendsRawData,
   vulnScanDataTransformed
@@ -69,71 +68,6 @@ const VulnerabilityScan: React.FC = () => {
   // >(undefined);
   const [vulnScanData, setVulnScanData] =
     useState<vulnScanDataTransformed>(InitialVSData);
-
-  const [topVulnHostsSeverityLevel, setTopVulnHostsSeverityLevel] =
-    useState('all');
-  const [topVulnHostsActiveLabel, setTopVulnHostsActiveLabel] = useState('all');
-
-  const [severityByProminenceVulnType, setSeverityByProminenceVulnType] =
-    useState('KEV');
-  const [severityByProminenceActiveLabel, setSeverityByProminenceActiveLabel] =
-    useState('KEV');
-
-  // const [topVulnDataType, setTopVulnDataType] = useState('All');
-  // const [topVulnActiveLabel, setTopVulnActiveLabel] = useState('All');
-  // const handleTopVulnChipClick = (label: string) => {
-  //   setTopVulnDataType(label);
-  //   setTopVulnActiveLabel(label);
-  // };
-  const handleTopVulnHostsChipClick = (label: string) => {
-    setTopVulnHostsSeverityLevel(label);
-    setTopVulnHostsActiveLabel(label);
-  };
-  const handleSevByProminenceChipClick = (label: string) => {
-    setSeverityByProminenceVulnType(label);
-    setSeverityByProminenceActiveLabel(label);
-  };
-
-  //Dynamically generate, format, and sort severity levels for Top Vuln Hosts' GraphChips from the data
-  const presentSeverityLevels = Array.from(
-    new Set(
-      vulnScanData.detectedHostsTop5VulnerableHosts.flatMap((item) =>
-        Object.entries(item)
-          .filter(
-            ([key, value]) =>
-              [
-                'all',
-                'lowSeverity',
-                'mediumSeverity',
-                'highSeverity',
-                'criticalSeverity'
-              ].includes(key) &&
-              typeof value === 'number' &&
-              value > 0
-          )
-          .map(([key]) => key)
-      )
-    )
-  )
-    .map((severity) => severity.replace('Severity', '').toLowerCase())
-    .sort((a: string, b: string) => {
-      const severityOrder = ['all', 'critical', 'high', 'medium', 'low'];
-      return severityOrder.indexOf(a) - severityOrder.indexOf(b);
-    });
-
-  const topVulnerableHostsGraphChips = presentSeverityLevels.map(
-    (severity) => ({
-      label: severity,
-      onClick: () => handleTopVulnHostsChipClick(severity)
-    })
-  );
-
-  const severityByProminenceGraphChips = ['KEV', 'Distinct', 'All'].map(
-    (type) => ({
-      label: type,
-      onClick: () => handleSevByProminenceChipClick(type)
-    })
-  );
 
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
@@ -293,83 +227,13 @@ const VulnerabilityScan: React.FC = () => {
         </Grid>
         <Grid container spacing={3}>
           <Grid size={{ xs: 12, lg: 6 }}>
-            <Box
-              border="1px solid"
-              borderRadius="4px"
-              borderColor="neutrals.light"
-              sx={{ px: 3, pb: 3, pt: 1 }}
-              minHeight="550px"
-              // maxWidth="538px"
-            >
-              <Grid
-                container
-                alignItems="center"
-                justifyContent="space-between"
-                paddingBottom={2}
-              >
-                <Grid size={{ xs: 12, lg: 7 }}>
-                  <InfoLabel
-                    label="Severity by Prominence"
-                    typographyVariant="h3"
-                    headingLevel="h3"
-                    tooltipContentJson={tooltipContentJson}
-                  />
-                </Grid>
-                <Grid size={{ xs: 12, lg: 5 }} justifyItems={'flex-end'}>
-                  <GraphChip
-                    data={severityByProminenceGraphChips}
-                    activeLabel={severityByProminenceActiveLabel}
-                  />
-                </Grid>
-              </Grid>
-              <SeverityByProminence
-                data={vulnScanData.severityByProminence}
-                activeVulnType={severityByProminenceVulnType}
-              />
-            </Box>
+            <SeverityByProminence data={vulnScanData.severityByProminence} />
           </Grid>
           <Grid size={{ xs: 12, lg: 6 }}>
             <TopVulnerabilities
               allData={vulnScanData.topVulnerabilities}
               kevData={vulnScanData.topKevVulnerabilities}
-              // activeVulnDataType={topVulnDataType}
             />
-            {/* <Box
-              border="1px solid"
-              borderRadius="4px"
-              borderColor="neutrals.light"
-              minHeight="550px"
-              sx={{ px: 3, pb: 1, pt: 1 }}
-              // maxWidth="538px"
-            >
-              <Grid
-                container
-                direction="row"
-                alignItems="center"
-                justifyContent="space-between"
-                paddingBottom={2}
-              >
-                <Grid size={{ xs: 12, lg: 9 }}>
-                  <InfoLabel
-                    label="Top Vulnerabilities by Occurrence"
-                    typographyVariant="h3"
-                    headingLevel="h3"
-                    tooltipContentJson={tooltipContentJson}
-                  />
-                </Grid>
-                <Grid size={{ xs: 12, lg: 3 }} justifyItems={'flex-end'}>
-                  <GraphChip
-                    data={topVulnerabilitiesGraphChips}
-                    activeLabel={topVulnActiveLabel}
-                  />
-                </Grid>
-              </Grid>
-              <TopVulnerabilities
-                allData={vulnScanData.topVulnerabilities}
-                kevData={vulnScanData.topKevVulnerabilities}
-                // activeVulnDataType={topVulnDataType}
-              />
-            </Box> */}
           </Grid>
         </Grid>
       </Box>
@@ -440,30 +304,9 @@ const VulnerabilityScan: React.FC = () => {
                 <KeyMetrics data={vulnScanData.detectedHostsKeyMetrics} />
               </Grid>
               <Grid size={{ xs: 12 }}>
-                <Box
-                  border="1px solid"
-                  borderRadius="4px"
-                  borderColor="neutrals.light"
-                  sx={{ px: 3, pb: 2.5, pt: 1 }}
-                  minHeight="435px"
-                >
-                  <InfoLabel
-                    label="Top Vulnerable Hosts"
-                    typographyVariant="h3"
-                    headingLevel="h3"
-                    tooltipContentJson={tooltipContentJson}
-                  />
-                  <Grid container direction="row" paddingBottom={2}>
-                    <GraphChip
-                      data={topVulnerableHostsGraphChips}
-                      activeLabel={topVulnHostsActiveLabel}
-                    />
-                  </Grid>
-                  <TopVulnerableHosts
-                    data={vulnScanData.detectedHostsTop5VulnerableHosts}
-                    severityLevel={topVulnHostsSeverityLevel}
-                  />
-                </Box>
+                <TopVulnerableHosts
+                  data={vulnScanData.detectedHostsTop5VulnerableHosts}
+                />
               </Grid>
             </Grid>
           </Grid>
